Added queue functionality to make a manual trigger

+++ src/MondayIntegrationSettings.cs
@@ -11,12 +11,12 @@ namespace monday_integration.src
 
         public string MondayBaseURL {get; private set;}
         public string MondayApiKey {get; private set;}
-        public int MondayAimsIntegrationBoardId {get; private set;}
+        public long MondayAimsIntegrationBoardId {get; private set;}
 
         public MondayIntegrationSettings(System.Collections.IDictionary dictionary) {
             foreach(var property in typeof(MondayIntegrationSettings).GetProperties()) {
-                if(property.PropertyType == typeof(int)) {
-                    property.SetValue(this, Int32.Parse(dictionary[property.Name].ToString()));
+                if(property.PropertyType == typeof(long)) {
+                    property.SetValue(this, long.Parse(dictionary[property.Name].ToString()));
                 } else {
                     property.SetValue(this, dictionary[property.Name]);
                 }
